Moving graphical isochrone and heat maps to simple.py
