+++ source/jormungandr/jormungandr/scenarios/simple.py
@@ -36,7 +36,6 @@ import navitiacommon.request_pb2 as request_pb2
 from jormungandr.interfaces.common import pb_odt_level
 from jormungandr.scenarios.utils import places_type, pt_object_type, add_link
 from jormungandr.scenarios.utils import build_pagination
-from jormungandr.scenarios.utils import updated_common_journey_request_with_default
 from jormungandr.exceptions import UnknownObject
 
 
@@ -47,46 +46,6 @@ def get_pb_data_freshness(request):
         return type_pb2.ADAPTED_SCHEDULE
     else:
         return type_pb2.BASE_SCHEDULE
-
-
-def isochrone_common(isochrone, request, instance, journey_req):
-
-    if request.get("origin"):
-        origin = journey_req.origin.add()
-        origin.place = request["origin"]
-        origin.access_duration = 0
-        journey_req.clockwise = True
-    if request.get("destination"):
-        destination = journey_req.destination.add()
-        destination.place = request["destination"]
-        destination.access_duration = 0
-        journey_req.clockwise = False
-    journey_req.datetimes.append(request["datetime"])
-    journey_req.wheelchair = request["wheelchair"] or False
-    journey_req.realtime_level = get_pb_data_freshness(request)
-    updated_common_journey_request_with_default(request, instance)
-    sn_params = journey_req.streetnetwork_params
-    sn_params.max_walking_duration_to_pt = request["max_walking_duration_to_pt"]
-    sn_params.max_bike_duration_to_pt = request["max_bike_duration_to_pt"]
-    sn_params.max_bss_duration_to_pt = request["max_bss_duration_to_pt"]
-    sn_params.max_car_duration_to_pt = request["max_car_duration_to_pt"]
-    sn_params.walking_speed = request["walking_speed"]
-    sn_params.bike_speed = request["bike_speed"]
-    sn_params.car_speed = request["car_speed"]
-    sn_params.bss_speed = request["bss_speed"]
-
-    journey_req.max_transfers = request["max_transfers"]
-    isochrone.origin_modes = request["origin_mode"]
-    isochrone.destination_modes = request["destination_mode"]
-    if "forbidden_uris[]" in request and request["forbidden_uris[]"]:
-        for forbidden_uri in request["forbidden_uris[]"]:
-            journey_req.forbidden_uris.append(forbidden_uri)
-    if "allowed_id[]" in request and request["allowed_id[]"]:
-        for allowed_id in request["allowed_id[]"]:
-            journey_req.allowed_id.append(allowed_id)
-
-    journey_req.streetnetwork_params.origin_mode = isochrone.origin_modes[0]
-    journey_req.streetnetwork_params.destination_mode = isochrone.destination_modes[0]
 
 
 class Scenario(object):
@@ -290,39 +249,6 @@ class Scenario(object):
         req.disable_disruption = request["disable_disruption"]
         resp = instance.send_and_receive(req)
         build_pagination(request, resp)
-        return resp
-
-    def graphical_isochrones(self, request, instance):
-        req = request_pb2.Request()
-        req._current_datetime = date_to_timestamp(request["_current_datetime"])
-        journey_req = req.isochrone.journeys_request
-        isochrone_common(self, request, instance, journey_req)
-        req.requested_api = type_pb2.graphical_isochrone
-        journey_req = req.isochrone.journeys_request
-
-        if request.get("max_duration"):
-            journey_req.max_duration = request["max_duration"]
-        else:
-            journey_req.max_duration = max(request["boundary_duration[]"], key=int)
-        if request.get("boundary_duration[]"):
-            for duration in sorted(request["boundary_duration[]"], key=int, reverse=True):
-                if request["min_duration"] < duration < journey_req.max_duration:
-                    req.isochrone.boundary_duration.append(duration)
-        req.isochrone.boundary_duration.insert(0, journey_req.max_duration)
-        req.isochrone.boundary_duration.append(request["min_duration"])
-        resp = instance.send_and_receive(req)
-        return resp
-
-    def heat_maps(self, request, instance):
-        req = request_pb2.Request()
-        req._current_datetime = date_to_timestamp(request["_current_datetime"])
-        journey_req = req.heat_map.journeys_request
-        isochrone_common(self, request, instance, journey_req)
-        req.requested_api = type_pb2.heat_map
-        max_resolution = 1000
-        req.heat_map.resolution = min(request["resolution"], max_resolution)
-        req.heat_map.journeys_request.max_duration = request["max_duration"]
-        resp = instance.send_and_receive(req)
         return resp
 
     def stop_areas(self, request, instance):
